refactor(http): move core tool builder to http-server (#852)
